Was off the vassar server for over 24 hours — in this time I fixed and fleshed out the shopFor() and adding to grocery list functionality, I added an editDialogue/alertDialogue for if a user adds an ingredient to the grocery list that’s already on there prompting them if they want to change quantity instead, and persisted this. fleshed out recipeDetailFragment, with tags. User can add tags and I made the full UI for that, including adding your own tag (still working on deleting dynamic tag). Went through all unit tests and updated all JavaDoc, as well as using unified addIngredient and typeText helper functions in all tests.

+++ astudio/app/src/main/java/com/example/pantrypalandroidprototype/model/RecipeBuilder.java
@@ -84,7 +84,7 @@ public class RecipeBuilder implements Serializable {
      * @param dietaryTags The dietary tags associated with the ingredient.
      * @return The current {@code RecipeBuilder} instance for method chaining.
      */
-    public RecipeBuilder addIngredient(String name, int quantity, String unit, Set<Ingredient.dietary_tags> dietaryTags) {
+    public RecipeBuilder addIngredient(String name, double quantity, String unit, Set<Ingredient.dietary_tags> dietaryTags) {
         // Convert dietary_tags to a set of strings
         Set<String> tagStrings = dietaryTags.stream()
                 .map(Enum::name)
